third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-drawImage-video.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>

description("Ensure correct behavior of drawImage with ImageBitmaps from video elements.");
window.jsTestIsAsync = true;

function jsWrapperClass(node)
{
    // returns the ClassName of node
    if (!node)
        return "[null]";
    var string = Object.prototype.toString.apply(node);

    // string will be of the form [object ClassName]
    return string.substr(8, string.length - 9);
}

function shouldBeType(expression, className)
{
    shouldBe("jsWrapperClass(" + expression + ")", "'" + className + "'");
}

function shouldBeClear(canvas, x, y) {
    // should be transparent black pixels
    var ctx = canvas.getContext('2d');
    d = ctx.getImageData(x, y, 1, 1).data;
    if (d[0] != 0 || d[1] != 0 || d[2] != 0 || d[3] != 0) {
        testFailed("This pixel must be clear, but it is" + d);
        finishJSTest();
    }
}

function clearContext(ctx, ctx1) {
    ctx.clearRect(0, 0, 500, 500);
    ctx1.clearRect(0, 0, 500, 500);
}

var bitmap;
var video;

var canvas = document.createElement("canvas");
canvas.setAttribute("width", "500");
canvas.setAttribute("height", "500");

var canvas1 = document.createElement("canvas");
canvas1.setAttribute("width", "500");
canvas1.setAttribute("height", "500");

video = document.createElement("video");
video.addEventListener("canplaythrough", videoLoaded, false);
video.src = "../../compositing/resources/video.ogv";

function videoLoaded() {
    var imageBitmaps = {};
    var p1 = createImageBitmap(video).then(function (image) { imageBitmaps.noCrop = image });
    var p2 = createImageBitmap(video, 0, 0, 100, 100).then(function (image) { imageBitmaps.crop = image });
    var p3 = createImageBitmap(video, 50, 50, 100, 100).then(function (image) { imageBitmaps.cropRight = image });
    var p4 = createImageBitmap(video, 100, 100, 100, 100).then(function (image) { imageBitmaps.cropCenter = image });
    var p5 = createImageBitmap(video, -100, -100, 600, 600).then(function (image) { imageBitmaps.overCrop = image });
    var p6 = createImageBitmap(video, 100, 100, 500, 500).then(function (image) { imageBitmaps.overCropRight = image });
    var p7 = createImageBitmap(video, 100, 100, -100, -100).then(function (image) { imageBitmaps.negativeCrop = image });
    var p8 = createImageBitmap(video, -300, -300, 300, 300).then(function (image) { imageBitmaps.empty = image });
    var p9 = createImageBitmap(video, 400, 300, 300, 300).then(function (image) { imageBitmaps.emptyTwo = image });
    var p10 = createImageBitmap(video, 0, 0, Math.pow(10, 6), Math.pow(10, 6)).then(function(image) { imageBitmaps.hugeCrop = image });
    Promise.all([p1, p2, p3, p4, p5, p6, p7, p8, p9]).then(function() {
        checkNoCrop(imageBitmaps.noCrop);
        checkCrop(imageBitmaps.crop, 0, 0, 100, 100);
        checkCrop(imageBitmaps.cropRight, 50, 50, 100, 100);
        checkCrop(imageBitmaps.cropCenter, 100, 100, 100, 100);
        checkOverCrop(imageBitmaps.overCrop);
        checkOverCropRight(imageBitmaps.overCropRight);
        checkCrop(imageBitmaps.negativeCrop, 0, 0, 100, 100);
        checkEmpty(imageBitmaps.empty);
        checkEmpty(imageBitmaps.emptyTwo);
        createImageBitmap(video, 0, 0, 0x8000, 0x8000).then(function() {
            testFailed('Creating a huge ImageBitmap is resolved unexpectedly.');
            finishJSTest();
         }, function() {
             testPassed('Creating a huge ImageBitmap is rejected as expected.');
             finishJSTest();
         });
    }, function() {
        testFailed("Promise was rejected.");
        finishJSTest();
    });
}

function compareTwoCanvases(sx, sy, sw, sh, ctx, ctx1)
{
    var data = ctx.getImageData(sx, sy, sw, sh).data;
    var data1 = ctx1.getImageData(sx, sy, sw, sh).data;
    var dataMatched = true;
    for (var i = 0; i < data.length; i++) {
        // data and data1 are strictly the same under software rendering, 
        // but significantly different on GPU.
        if (Math.abs(data[i] - data1[i]) > 80) {
            dataMatched = false;
            break;
        }
    }
    if (!dataMatched) {
        testFailed("The imageBitmap doesn't match the original video.");
        finishJSTest();
    }
}

function checkNoCrop(imageBitmap) {
    bitmap = imageBitmap;
    shouldBeType("bitmap", "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // should be drawn to (0, 0), (352, 288)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    ctx1.drawImage(video, 0, 0);
    compareTwoCanvases(0, 0, 352, 288, ctx, ctx1);

    // shrunk to (0, 0), (100, 100)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0, 100, 100);
    ctx1.drawImage(video, 0, 0, 100, 100);
    compareTwoCanvases(0, 0, 100, 100, ctx, ctx1);

    // shrunk to (100, 100), (250, 200)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 100, 100, 150, 100);
    ctx1.drawImage(video, 100, 100, 150, 100);
    compareTwoCanvases(100, 100, 150, 100, ctx, ctx1); 

    // black should be drawn to (100, 100), (200, 200)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 100, 100, 100, 100, 100, 100, 100, 100);
    ctx1.drawImage(video, 100, 100, 100, 100, 100, 100, 100, 100);
    compareTwoCanvases(100, 100, 100, 100, ctx, ctx1);

    // nothing should be drawn
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 400, 300, 200, 200, 100, 100, 100, 100);
    shouldBeClear(canvas, 10, 10);
    shouldBeClear(canvas, 80, 80);
    shouldBeClear(canvas, 150, 150);
    shouldBeClear(canvas, 210, 210);
}

function checkCrop(imageBitmap, sx, sy, sw, sh) {
    bitmap = imageBitmap;
    shouldBeType("bitmap", "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // should be drawn to (0, 0), (100, 100)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    ctx1.drawImage(video, sx, sy, sw, sh, 0, 0, sw, sh);
    compareTwoCanvases(0, 0, 100, 100, ctx, ctx1);
}

function checkOverCrop(imageBitmap) {
    bitmap = imageBitmap;
    shouldBeType("bitmap", "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // should be drawn to (100, 100), (452, 388)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    ctx1.drawImage(video, 100, 100, 352, 288);
    compareTwoCanvases(0, 0, 500, 500, ctx, ctx1);

    // should be drawn to (50, 50), (226, 194)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0, 300, 300);
    ctx1.drawImage(video, 0, 0, 352, 288, 50, 50, 176, 144);
    compareTwoCanvases(0, 0, 300, 300, ctx, ctx1);
}

function checkOverCropRight(imageBitmap) {
    bitmap = imageBitmap;
    shouldBeType("bitmap", "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // should be drawn to (0, 0), (252, 188)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    ctx1.drawImage(video, 100, 100, 252, 188, 0, 0, 252, 188);
    compareTwoCanvases(0, 0, 252, 188, ctx, ctx1);
}

function checkEmpty(imageBitmap) {
    bitmap = imageBitmap;
    shouldBeType("bitmap", "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // nothing should be drawn
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    shouldBeClear(canvas, 10, 10);
    shouldBeClear(canvas, 90, 90);
    shouldBeClear(canvas, 110, 110);
    shouldBeClear(canvas, 210, 210);
}

</script>
</body>
</html>
